♻️ move createHttpRequest to @datadog/js-core/transport
fetch and sendBeacon are browser-specific APIs; they become injectable
strategies so createHttpRequest stays free of browser globals.

browser-core/transport/httpRequest.ts keeps fetchStrategy and
sendBeaconStrategy and wraps the js-core function by injecting them.
The public signature (endpointBuilders, reportError, bytesLimit?) is
unchanged for all existing callers.
