+++ packages/js-core/src/entries/transport.ts
@@ -12,8 +12,12 @@ export { createEndpointBuilder, createReplicaEndpointBuilder, buildEndpointUrl }
 export type { Encoder, EncoderResult } from '../transport/encoder'
 export { createIdentityEncoder } from '../transport/encoder'
 export type { Payload, HttpResponse, BandwidthStats, HttpRequestEvent } from '../transport/payload'
+export type { HttpRequest, SendStrategy, SendOnExitStrategy } from '../transport/httpRequest'
+export { createHttpRequest } from '../transport/httpRequest'
+export type { Batch } from '../transport/batch'
+export { createBatch, MESSAGE_BYTES_LIMIT } from '../transport/batch'
 export { RECOMMENDED_REQUEST_BYTES_LIMIT } from '../transport/payload'
-export { PageExitReason, createFlushController, FLUSH_DURATION_LIMIT, MESSAGES_LIMIT } from '../transport/flushController'
+export { PageExitReason, isPageExitReason, createFlushController, FLUSH_DURATION_LIMIT, MESSAGES_LIMIT } from '../transport/flushController'
 export type { PageMayExitEvent, UrgentFlushReason, FlushReason, FlushEvent, FlushController } from '../transport/flushController'
 export {
   sendWithRetryStrategy,
